requirements.txt
mysql-connector-python
# Tkinter viene incluido en la mayoría de instalaciones de Python, pero si es necesario:
tk 